+++ scripts/macros/classFeatures/artificer/battleSmith/steelDefender.js
@@ -33,6 +33,7 @@ async function item({speaker, actor, token, character, item, args, scope, workfl
     if (arcaneJoltItem) arcaneJoltData.system.uses = arcaneJoltItem.system.uses;
     let hpValue = (2 + chris.getSpellMod(workflow.item)) + (artificerLevel * 5);
     let name = chris.getConfiguration(workflow.item, 'name') ?? 'Steel Defender';
+    if (name === '') name = 'Steel Defender';
     let updates = {
         'actor': {
             'name': name,