+++ scripts/macros/classFeatures/artificer/infusions/homunculusServant.js
@@ -23,6 +23,7 @@ export async function homunculusServant({speaker, actor, token, character, item,
     let hpFormula = artificerLevel + 'd4 + ' + (1 + chris.getSpellMod(workflow.item));
     let hpValue = await new Roll(hpFormula).evaluate({async: true});
     let name = chris.getConfiguration(workflow.item, 'name') ?? 'Homunculus Servant';
+    if (name === '') name = 'Homunculus Servant';
     let updates = {
         'actor': {
             'name': name,