+++ scripts/macros/classFeatures/druid/circleOfWildfire/summonWildfireSpirit.js
@@ -34,6 +34,7 @@ async function item({speaker, actor, token, character, item, args, scope, workfl
     dodgeData.system.description.value = chris.getItemDescription('CPR - Descriptions', 'Dodge');
     let hpFormula = 5 + (druidLevel * 5);
     let name = chris.getConfiguration(workflow.item, 'name') ?? 'Wildfire Spirit';
+    if (name === '') name = 'Wildfire Spirit';
     let updates = {
         'actor': {
             'name': name,