+++ scripts/macros/spells/summonBeast.js
@@ -16,6 +16,7 @@ export async function summonBeast({speaker, actor, token, character, item, args,
     maulData.system.damage.parts[0][0] += ' + ' + workflow.castData.castLevel;
     let hpFormula;
     let name = chris.getConfiguration(workflow.item, 'name-' + selection) ?? 'Bestial Spirit (' + selection + ')';
+    if (name === '') name = 'Bestial Spirit (' + selection + ')';
     let updates = {
         'actor': {
             'name': name,