+++ scripts/macros/spells/summonDraconicSpirit.js
@@ -21,6 +21,7 @@ export async function summonDraconicSpirit({speaker, actor, token, character, it
     breathData.system.save.dc = chris.getSpellDC(workflow.item);
     let hpFormula = 50 + ((workflow.castData.castLevel - 5) * 10);
     let name = chris.getConfiguration(workflow.item, 'name-' + selection) ?? 'Draconic Spirit (' + selection + ')';
+    if (name === '') name = 'Draconic Spirit (' + selection + ')';
     let updates = {
         'actor': {
             'name': name,