+++ scripts/macros/spells/summonElemental.js
@@ -16,6 +16,7 @@ export async function summonElemental({speaker, actor, token, character, item, a
     slamData.system.damage.parts[0][0] += ' + ' + workflow.castData.castLevel;
     let hpFormula = 50 + ((workflow.castData.castLevel - 4) * 10);
     let name = chris.getConfiguration(workflow.item, 'name-' + selection) ?? 'Elemental Spirit (' + selection + ')';
+    if (name === '') name = 'Elemental Spirit (' + selection + ')';
     let updates = {
         'actor': {
             'name': name,