+++ scripts/macros/spells/summonShadowspawn.js
@@ -20,6 +20,7 @@ export async function summonShadowspawn({speaker, actor, token, character, item,
     dreadfulScreamData.system.save.dc = chris.getSpellDC(workflow.item);
     let hpFormula = 35 + ((workflow.castData.castLevel - 3) * 15);
     let name = chris.getConfiguration(workflow.item, 'name-' + selection) ?? 'Shadow Spirit (' + selection + ')';
+    if (name === '') name = 'Shadow Spirit (' + selection + ')';
     let updates = {
         'actor': {
             'name': name,