+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # 0.5.37 Release:
+## New Content:
+- Spiritual Weapon
 ## Update Notes:
 - Added configuration option for custom names for summon spells and class features.
 ## Bug Fixes: